Change Movie id from Short to Int to match DDBB type

+++ src/main/scala/io/movies/algebras/Repository.scala
@@ -5,7 +5,7 @@ import io.movies.model.{Movie, RegisteredMovie}
 trait Repository[F[_]] {
   def addMovie(movie: Movie): F[Int]
   def getMovies: F[List[RegisteredMovie]]
-  def getMovieById(id: Short): F[Option[RegisteredMovie]]
+  def getMovieById(id: Int): F[Option[RegisteredMovie]]
   def createMovieTable: F[Int]
   def dropMovieTable: F[Int]
 }
